requirements.txt
Django==3.0.3
djangorestframework
